configs/kueue/dev/kueuer/pyproject.toml
[project]
name = "kueuer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Shiny Brar", email = "shiny.brar@nrc-cnrc.gc.ca" }
]
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=24.1.0",
    "httpx>=0.28.1",
    "ipython>=8.18.1",
    "kubernetes>=32.0.1",
    "logfire[httpx]>=3.7.1",
    "nest-asyncio>=1.6.0",
    "pyyaml>=6.0.2",
    "skaha",
    "typer>=0.15.2",
]
scripts = { kue-stress = "kueuer.jobs.launcher:app" }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
skaha = { git = "https://github.com/shinybrar/skaha.git", branch = "CADC-14066-scaling-issues" }

[tool.ruff.lint]
select = ["E", "F", "W", "C", "N", "B", "I"]
ignore = ["B904", "B028"]
extend-ignore = ["D"]

[tool.ruff.lint.isort]
combine-as-imports = true
